backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
langchain-openai
langchain-core
openpyxl
pydantic